Join data frame to species group definitions

This function will join a data.frame or tibble object to the species group definitions using the species code. It can be used following the calculation of overlap or captures per species.

Usage

join_species_groups(object, lookup, ...)

# S3 method for class 'tbl_df'
join_species_groups(object, lookup, ...)

# S3 method for class 'data.frame'
join_species_groups(object, lookup, ...)

Arguments

object

A data.frame or tibble object with a column containing the species codes.

lookup

A data.frame or tibble object providing a lookup table with species groups definitions.

...

Arguments to be passed to left_join.

Note

This function uses left_join to assign species group definitions to the object argument. Species will not be included in the output if they are not present in the input.